+++ benches/benches/encode-geyser/block_meta.rs
@@ -1,14 +1,8 @@
 use {
-    criterion::{BatchSize, Criterion},
-    prost::Message,
-    prost_types::Timestamp,
+    criterion::Criterion,
     richat_benches::fixtures::generate_block_metas,
     richat_plugin_agave::protobuf::{ProtobufEncoder, ProtobufMessage},
-    richat_proto::plugin::{
-        filter::message::{FilteredUpdate, FilteredUpdateFilters, FilteredUpdateOneof},
-        message::MessageBlockMeta,
-    },
-    std::{hint::black_box, sync::Arc, time::SystemTime},
+    std::{hint::black_box, time::SystemTime},
 };
 
 pub fn bench_encode_block_metas(criterion: &mut Criterion) {
@@ -17,12 +11,6 @@ pub fn bench_encode_block_metas(criterion: &mut Criterion) {
     let blocks_meta_replica = blocks_meta
         .iter()
         .map(|b| b.to_replica())
-        .collect::<Vec<_>>();
-
-    let blocks_meta_grpc = blocks_meta_replica
-        .iter()
-        .map(MessageBlockMeta::from_geyser)
-        .map(Arc::new)
         .collect::<Vec<_>>();
 
     criterion
@@ -57,49 +45,6 @@ pub fn bench_encode_block_metas(criterion: &mut Criterion) {
                         }
                     })
                 })
-            },
-        )
-        .bench_with_input(
-            "dragons-mouth/encoding-only",
-            &blocks_meta_grpc,
-            |criterion, messages| {
-                let created_at = Timestamp::from(SystemTime::now());
-                criterion.iter_batched(
-                    || messages.to_owned(),
-                    |messages| {
-                        #[allow(clippy::unit_arg)]
-                        black_box({
-                            for message in messages {
-                                let update = FilteredUpdate {
-                                    filters: FilteredUpdateFilters::new(),
-                                    message: FilteredUpdateOneof::block_meta(message),
-                                    created_at,
-                                };
-                                update.encode_to_vec();
-                            }
-                        })
-                    },
-                    BatchSize::LargeInput,
-                );
-            },
-        )
-        .bench_with_input(
-            "dragons-mouth/full-pipeline",
-            &blocks_meta_replica,
-            |criterion, block_metas| {
-                let created_at = Timestamp::from(SystemTime::now());
-                criterion.iter(|| {
-                    #[allow(clippy::unit_arg)]
-                    black_box(for blockinfo in block_metas {
-                        let message = MessageBlockMeta::from_geyser(blockinfo);
-                        let update = FilteredUpdate {
-                            filters: FilteredUpdateFilters::new(),
-                            message: FilteredUpdateOneof::block_meta(Arc::new(message)),
-                            created_at,
-                        };
-                        update.encode_to_vec();
-                    })
-                });
             },
         );
 }