+++ benches/benches/encode-geyser/transaction.rs
@@ -1,13 +1,7 @@
 use {
     criterion::Criterion,
-    prost::Message,
-    prost_types::Timestamp,
     richat_benches::fixtures::generate_transactions,
     richat_plugin_agave::protobuf::{ProtobufEncoder, ProtobufMessage},
-    richat_proto::plugin::{
-        filter::message::{FilteredUpdate, FilteredUpdateFilters, FilteredUpdateOneof},
-        message::MessageTransaction,
-    },
     std::{hint::black_box, time::SystemTime},
 };
 
@@ -17,11 +11,6 @@ pub fn bench_encode_transactions(criterion: &mut Criterion) {
     let transactions_replica = transactions
         .iter()
         .map(|tx| tx.to_replica())
-        .collect::<Vec<_>>();
-
-    let transactions_grpc = transactions_replica
-        .iter()
-        .map(|(slot, transaction)| MessageTransaction::from_geyser(transaction, *slot))
         .collect::<Vec<_>>();
 
     criterion
@@ -59,47 +48,6 @@ pub fn bench_encode_transactions(criterion: &mut Criterion) {
                                 transaction,
                             };
                             message.encode_with_timestamp(ProtobufEncoder::Raw, created_at);
-                        }
-                    })
-                });
-            },
-        )
-        .bench_with_input(
-            "dragons-mouth/encoding-only",
-            &transactions_grpc,
-            |criterion, transaction_messages| {
-                let created_at = Timestamp::from(SystemTime::now());
-                criterion.iter(|| {
-                    #[allow(clippy::unit_arg)]
-                    black_box({
-                        for message in transaction_messages {
-                            let update = FilteredUpdate {
-                                filters: FilteredUpdateFilters::new(),
-                                message: FilteredUpdateOneof::transaction(message),
-                                created_at,
-                            };
-                            update.encode_to_vec();
-                        }
-                    })
-                });
-            },
-        )
-        .bench_with_input(
-            "dragons-mouth/full-pipeline",
-            &transactions_replica,
-            |criterion, transactions| {
-                let created_at = Timestamp::from(SystemTime::now());
-                criterion.iter(|| {
-                    #[allow(clippy::unit_arg)]
-                    black_box({
-                        for (slot, transaction) in transactions {
-                            let message = MessageTransaction::from_geyser(transaction, *slot);
-                            let update = FilteredUpdate {
-                                filters: FilteredUpdateFilters::new(),
-                                message: FilteredUpdateOneof::transaction(&message),
-                                created_at,
-                            };
-                            update.encode_to_vec();
                         }
                     })
                 });