+++ richat/src/pubsub/notification.rs
@@ -10,9 +10,8 @@ use {
     richat_shared::five8::signature_encode,
     serde::Serialize,
     solana_clock::Slot,
-    solana_message::{VersionedMessage, v0::LoadedMessage},
+    solana_message_v3::{VersionedMessage, v0::LoadedMessage},
     solana_rpc_client_api::response::{Response as RpcResponse, RpcResponseContext},
-    solana_transaction::versioned::TransactionVersion,
     solana_transaction_status::{
         BlockEncodingOptions, EncodeError, EncodedTransaction, EncodedTransactionWithStatusMeta,
         TransactionDetails, TransactionStatusMeta, UiAccountsList, UiConfirmedBlock,
@@ -20,6 +19,7 @@ use {
         option_serializer::OptionSerializer,
         parse_accounts::{parse_legacy_message_accounts, parse_v0_message_accounts},
     },
+    solana_transaction_v3::versioned::TransactionVersion,
     std::{
         collections::VecDeque,
         sync::{Arc, Weak},